Updated NStarAlignment code to remove the need to reference ASCOM. Updated tests to accept new parameters.

+++ NStarAlignment/DataTypes/SphericalCoordinate.cs
@@ -7,7 +7,7 @@ namespace NStarAlignment.DataTypes
     {
         public Angle X;
         public Angle Y;
-        public double R;
+        public int R;
 
         #region Operator overloads ...
         /// <summary>
